chat_with_nested_images/requirements.txt
langchain-community
langchain
langchain-openai
pymilvus
numpy 
streamlit
pandas
torchvision
torch
openpyxl
python-dotenv